Cleanup bumblebee integration
