+++ ftw/file/configure.zcml
@@ -41,7 +41,10 @@
         description="Installs the ftw.file package with bumblebee integration"
         provides="Products.GenericSetup.interfaces.EXTENSION"
         />
+    <include package=".bumblebee" />
   </configure>
+
+
   <profilehook:hook
       profile="ftw.file:default"
       handler=".hooks.installed"
