+++ ftw/file/tests/test_file_preview_actions_collector.py
@@ -1,6 +1,6 @@
 from ftw.builder import Builder
 from ftw.builder import create
-from ftw.file.interfaces import IFilePreviewActionsCollector
+from ftw.file.bumblebee.interfaces import IFilePreviewActionsCollector
 from ftw.file.testing import FTW_FILE_BUMBLEBEE_FUNCTIONAL_TESTING
 from plone.app.testing import logout
 from plone.app.testing import setRoles
